+++ src/rest_api/classes/gene.clj
@@ -6,7 +6,7 @@
    [rest-api.classes.gene.widgets.sequences :as sequences]
    [rest-api.classes.gene.widgets.genetics :as genetics]
    [rest-api.classes.gene.widgets.history :as history]
-   [rest-api.classes.gene.widgets.interactions :as interactions]
+   [rest-api.classes.interaction.core :as interactions]
    [rest-api.classes.gene.widgets.mapping-data :as mapping-data]
    [rest-api.classes.gene.widgets.ontology :as gene-ontology]
    [rest-api.classes.gene.widgets.references :as references]