+++ CHANGES.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.6.0] - 2018--07-05
+- Adding endpoints for interaction widgets
+- Adding detailed disease model table
+
 ## [0.5.2] - 2018-06-08
 - Fixing issue with variation